Calculator project started
Wrote functions for basic operations on calculator, added GUI buttons for numbers and operators
Numbers 0-9 and operators except = and AC are working as intended.
Need to address logic on calculation related to adding numbers to display (see line 32 in js)
Functionality for AC and = to be added to complete script portion.

CSS styling to follow scipting working as intended.